% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_latest.R
\name{find_latest}
\alias{find_latest}
\title{Find latest version of a file}
\usage{
find_latest(
  pattern,
  where = getwd(),
  date_parser = lubridate::ymd,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{pattern}{a character string indicating the name of the file to look for, or
regular expression to be matched against file names}

\item{where}{the directory in which to look for the file; defaults to the
current directory}

\item{date_parser}{a function to extract dates from character strings; we
recommend using \code{lubridate}'s function there, e.g. \code{ymd} (any date format
with year, month, day) or \code{dmy} (any format using day, month, year);
defaults to \code{lubridate::ymd}}

\item{quiet}{a \code{logical} indicating if messages should be displayed (\code{TRUE},
default), or not (\code{FALSE})}

\item{...}{further arguments passed to \url{list.files}}
}
\description{
If some file names include dates, this function will find the latest version
of a given file. This can be handy e.g. for finding the latest data update
automatically. \code{find_latest} scans recursively a directory, looking for file
names matching a given pattern (or regular expression) and including
dates. Dates are extracted using \url{extract_date}, using a
user-provided parser (see details). The full path to the most recent file
will be returned.
}
\details{
The date parser used by default is \url{lubridate::ymd},
so that any date specified as year/month/day (regardless of the separator
used) should work. One exception is that only dates provided as numbers
can be used e.g. "1982.02.04" and "82/2/4" are okay but not
"1982 Feb 04". For dates provided in a different order, check other
parsers implemented in \code{lubridate}, e.g. \code{lubridate::dmy} for
day-month-year formats, or \code{lubridate::ydm} for year-day-month. Note that
parsers for date-time objects can also be used, but in any case a \code{Date}
object will be returned.
}
\examples{
 

if (require(reportfactory)) {

## create random factory and toy files
odir <- getwd()
f <- new_factory(path = tempdir())

file.create(file.path("data", "linelist_2020-10-01.xlsx"))
file.create(file.path("data", "linelist_2020-10-12.csv"))
file.create(file.path("data", "linelist.xlsx"))
file.create(file.path("data", "contacts.xlsx"))
file.create(file.path("data", "death_linelist_2020-10-13.xlsx"))

## find the latest data with 'linelist' in the name; note that this
## matches both 'linelist' and 'death_linelist' files
find_latest("linelist")

## same, but this time files starting with 'linelist', i.e. excluding
## 'death_linelist'
find_latest("^linelist")

## this returns NULL
find_latest("foobar")

## cleanup
unlink(f, recursive = TRUE)
setwd(odir)
}
}
\seealso{
\url{extract_date} to extract dates from \code{character}
strings; see the \code{lubridate} \href{https://lubridate.tidyverse.org/}{website}
and
\href{https://rawgit.com/rstudio/cheatsheets/master/lubridate.pdf}{cheatsheet}
for other parsers
}
\author{
Thibaut Jombart
}
