% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scripts.R
\name{load_scripts}
\alias{load_scripts}
\title{Load R scripts within factory folder}
\usage{
load_scripts(factory = ".", folder = "scripts", quiet = FALSE)
}
\arguments{
\item{factory}{The path to the report factory or a folder within the desired
factory. Defaults to the current directory.}

\item{folder}{The folder (relative to the factory) from which to source \code{.R}
files; defaults to \verb{scripts/}.}

\item{quiet}{A logical indicating whether messages should be displayed to the
console (\code{TRUE}, default), or not.}
}
\description{
This helper function will load all \code{.R} scripts within a specified factory
and folder. All scripts are loaded within the calling environment and in
alphanumeric order.
}
\examples{

if (require(reportfactory)) {

# create factory
f <- new_factory(path = tempdir(), move_in = FALSE)

# create .R files in scripts/

## scripts are loaded in alphanumeric order, so "b.R" is loaded after "a.R" and
## can use its content

cat("addone <- function(x) x + 1", file = file.path(f, "scripts", "addone.R"))
cat("a <- 2L", file = file.path(f, "scripts", "a.R"))
cat("b <- 2L + a", file = file.path(f, "scripts", "b.R"))

## load scripts, check that 'a', 'b', and 'addone' exist
load_scripts(f)
a
b
addone(10)
addone(c(a, b))

# cleanup
unlink(f, recursive = TRUE)
}
}
